Move separate files into functions in index
