+++ public/profile/index.php
@@ -2,7 +2,6 @@
 include __DIR__ . ("/../auth_session.php");
 $user = $_GET['for'] ?? $_SESSION['username'];
 $for = $_GET['for'] ?? "''";
-require_once (__DIR__ . '/../../src/Ephew-internals/functions.php');
 $SQL_comm_USER = GetSQLCreds('username');
 $SQL_comm_PASS = GetSQLCreds('password');
 // Check connection
@@ -14,7 +13,7 @@ if (mysqli_connect_errno()) {
 if ((($user === $for)) and (!($user === $_SESSION['username']))) {
     $pagetitle = $user . "'s profile";
     $usedefaultsidebar = "false";
-    include __DIR__ . ("/../../src/Ephew-internals/unifiedheader.php");
+    unifiedheader();
 ?>
     <button class="openbtn" onclick="openNav()">☰</button>
     <div class="sidebar" id="mySidebar"><a href="javascript:void(0)" class="closebtn" onclick="closeNav()">×</a>
@@ -32,7 +31,7 @@ if ((($user === $for)) and (!($user === $_SESSION['username']))) {
 if ((($user === $_SESSION['username']))) {
     $pagetitle = "Your profile";
     $usedefaultsidebar = "false";
-    include __DIR__ . ("/../../src/Ephew-internals/unifiedheader.php");
+    unifiedheader();
     ?>
         <button class="openbtn" onclick="openNav()">☰</button>
         <div class="sidebar" id="mySidebar"><a href="javascript:void(0)" class="closebtn" onclick="closeNav()">×</a>
@@ -78,5 +77,5 @@ echo "</ul>";
 
         </div>
         <?php
-        include __DIR__ . ("/../../src/Ephew-internals/unifiedfooter.php");
+        unifiedfooter();
         ?>
